gui/requirements.txt
tk
customtkinter
Pillow